feat(email): merge producer and consumer into single Worker
- Replace packages/email-producer + packages/email-consumer with packages/email
- Single wrangler.jsonc with both queue producers and consumers
- Single Worker export with fetch (HTTP /enqueue) and queue handlers
- Extract shared EmailJob type; share TestD1Database via test helper
- Consolidate mise.toml email tasks: email:check and email:test
- Consolidate mise-tasks: check / test / clean / deploy all updated
- Consolidate CI jobs: email-check and email-test (4 jobs → 2)
