+++ .github/workflows/ci.yml
@@ -89,7 +89,7 @@ jobs:
       - name: Test
         run: mise run frontend:test
 
-  email-producer-check:
+  email-check:
     runs-on: ubuntu-latest
     steps:
       - uses: actions/checkout@v6
@@ -100,9 +100,9 @@ jobs:
         run: mise run frontend:install
 
       - name: Check
-        run: mise run email:producer:check
+        run: mise run email:check
 
-  email-producer-test:
+  email-test:
     runs-on: ubuntu-latest
     steps:
       - uses: actions/checkout@v6
@@ -113,33 +113,7 @@ jobs:
         run: mise run frontend:install
 
       - name: Test
-        run: mise run email:producer:test
-
-  email-consumer-check:
-    runs-on: ubuntu-latest
-    steps:
-      - uses: actions/checkout@v6
-
-      - uses: jdx/mise-action@v4
-
-      - name: Install JS dependencies
-        run: mise run frontend:install
-
-      - name: Check
-        run: mise run email:consumer:check
-
-  email-consumer-test:
-    runs-on: ubuntu-latest
-    steps:
-      - uses: actions/checkout@v6
-
-      - uses: jdx/mise-action@v4
-
-      - name: Install JS dependencies
-        run: mise run frontend:install
-
-      - name: Test
-        run: mise run email:consumer:test
+        run: mise run email:test
 
   backend-check:
     runs-on: ubuntu-latest
