+++ mise-tasks/email/deploy
@@ -1,0 +1,5 @@
+#!/usr/bin/env bash
+#MISE dir="packages/email"
+#MISE description="Deploy email Worker"
+
+pnpm run deploy
