--- packages/email-consumer/package.json
@@ -1,29 +1,0 @@
-{
-  "name": "@portaldots/email-consumer",
-  "private": true,
-  "type": "module",
-  "scripts": {
-    "dev:high": "wrangler dev --env high --port 8788 --inspector-port 9230 --persist-to .wrangler/state-high",
-    "dev:normal": "wrangler dev --env normal --port 8789 --inspector-port 9231 --persist-to .wrangler/state-normal",
-    "deploy:high": "wrangler deploy --env high",
-    "deploy:normal": "wrangler deploy --env normal",
-    "typecheck": "tsgo --noEmit",
-    "typecheck:compat": "tsc --noEmit",
-    "lint": "oxlint --type-aware .",
-    "format": "oxfmt .",
-    "format:check": "oxfmt --check .",
-    "ci:check": "$npm_execpath run typecheck && $npm_execpath run lint && $npm_execpath run format:check",
-    "test": "vitest run"
-  },
-  "dependencies": {
-    "hono": "4.12.22",
-    "hono-email": "0.4.0"
-  },
-  "devDependencies": {
-    "@cloudflare/workers-types": "4.20260522.1",
-    "@typescript/native-preview": "7.0.0-dev.20260314.1",
-    "typescript": "6.0.3",
-    "vitest": "4.1.7",
-    "wrangler": "4.94.0"
-  }
-}
