--- packages/email-consumer/wrangler.jsonc
@@ -1,84 +1,0 @@
-{
-  "$schema": "node_modules/wrangler/config-schema.json",
-  "name": "email-consumer",
-  "main": "src/index.ts",
-  "compatibility_date": "2026-05-02",
-  "d1_databases": [
-    {
-      "binding": "DB",
-      "database_name": "email-idempotency",
-      "database_id": "00000000-0000-0000-0000-000000000000"
-    }
-  ],
-  "queues": {
-    "consumers": [
-      {
-        "queue": "email-high",
-        "max_batch_size": 1,
-        "max_batch_timeout": 1,
-        "max_retries": 5
-      },
-      {
-        "queue": "email-normal",
-        "max_batch_size": 10,
-        "max_batch_timeout": 30,
-        "max_retries": 5
-      }
-    ]
-  },
-  "send_email": [
-    {
-      "name": "EMAIL"
-    }
-  ],
-  "env": {
-    "high": {
-      "queues": {
-        "consumers": [
-          {
-            "queue": "email-high",
-            "max_batch_size": 1,
-            "max_batch_timeout": 1,
-            "max_retries": 5
-          }
-        ]
-      },
-      "d1_databases": [
-        {
-          "binding": "DB",
-          "database_name": "email-idempotency",
-          "database_id": "00000000-0000-0000-0000-000000000000"
-        }
-      ],
-      "send_email": [
-        {
-          "name": "EMAIL"
-        }
-      ]
-    },
-    "normal": {
-      "queues": {
-        "consumers": [
-          {
-            "queue": "email-normal",
-            "max_batch_size": 10,
-            "max_batch_timeout": 30,
-            "max_retries": 5
-          }
-        ]
-      },
-      "d1_databases": [
-        {
-          "binding": "DB",
-          "database_name": "email-idempotency",
-          "database_id": "00000000-0000-0000-0000-000000000000"
-        }
-      ],
-      "send_email": [
-        {
-          "name": "EMAIL"
-        }
-      ]
-    }
-  }
-}
