+++ packages/email/.dev.vars.example
