+++ packages/email/.gitignore
