+++ packages/email/package.json
@@ -1,10 +1,10 @@
 {
-  "name": "@portaldots/email-producer",
+  "name": "@portaldots/email",
   "private": true,
   "type": "module",
   "scripts": {
     "dev": "wrangler dev --port 8787 --inspector-port 9229",
-    "dev:local-stack": "wrangler d1 migrations apply email-idempotency --local --persist-to .wrangler/state && wrangler dev -c wrangler.jsonc -c ../email-consumer/wrangler.jsonc --persist-to .wrangler/state --port 8787 --inspector-port 9229",
+    "dev:local-stack": "wrangler d1 migrations apply email-idempotency --local --persist-to .wrangler/state && wrangler dev --persist-to .wrangler/state --port 8787 --inspector-port 9229",
     "deploy": "wrangler deploy",
     "typecheck": "tsgo --noEmit",
     "typecheck:compat": "tsc --noEmit",
@@ -16,6 +16,7 @@
   },
   "dependencies": {
     "hono": "4.12.22",
+    "hono-email": "0.4.0",
     "zod": "4.4.3"
   },
   "devDependencies": {
