+++ packages/email/src/templates/index.ts
