+++ packages/email/src/templates/registration-verify.tsx
