+++ packages/email/src/templates/staff-auth-notice.tsx
