+++ packages/email/src/templates/styles.tsx
