+++ packages/email/src/tests/consumer.test.ts
@@ -1,149 +1,18 @@
 import { describe, it, expect, vi, beforeEach } from 'vitest'
 
+vi.mock('../templates', () => ({
+  renderTemplate: vi.fn()
+}))
+
+import { queueHandler } from '../consumer'
+import { renderTemplate } from '../templates'
+import { TestD1Database } from './helpers/d1'
+
 interface MessageLike {
   body: unknown
   ack: () => void
   retry: () => void
 }
-
-interface D1RunResult {
-  meta: { changes: number }
-  results: unknown[]
-  success: true
-}
-
-class TestD1Statement {
-  private values: unknown[] = []
-
-  constructor(
-    private readonly query: string,
-    private readonly db: TestD1Database
-  ) {}
-
-  bind(...values: unknown[]): TestD1Statement {
-    this.values = values
-    return this
-  }
-
-  async run(): Promise<D1RunResult> {
-    return this.db.run(this.query, this.values)
-  }
-
-  async first<T>(): Promise<T | null> {
-    return this.db.first<T>(this.query, this.values)
-  }
-}
-
-class TestD1Database {
-  readonly jobs = new Map<string, { status: string }>()
-  readonly chunks = new Map<string, { jobId: string; status: string; updatedAt: string }>()
-  failSentUpdate = false
-
-  prepare(query: string): TestD1Statement {
-    return new TestD1Statement(query, this)
-  }
-
-  async batch(statements: TestD1Statement[]): Promise<D1RunResult[]> {
-    return Promise.all(statements.map((statement) => statement.run()))
-  }
-
-  async run(query: string, values: unknown[]): Promise<D1RunResult> {
-    if (query.includes('INSERT OR IGNORE INTO email_jobs')) {
-      const jobId = String(values[0])
-      if (!this.jobs.has(jobId)) {
-        this.jobs.set(jobId, { status: 'queued' })
-        return runResult(1)
-      }
-      return runResult(0)
-    }
-    if (query.includes('INSERT OR IGNORE INTO email_job_chunks')) {
-      const messageId = String(values[0])
-      if (!this.chunks.has(messageId)) {
-        this.chunks.set(messageId, { jobId: String(values[1]), status: 'queued', updatedAt: String(values[6]) })
-        return runResult(1)
-      }
-      return runResult(0)
-    }
-    if (query.includes("SET status = 'processing'")) {
-      const messageId = String(values[1])
-      const chunk = this.chunks.get(messageId)
-      if (!chunk) return runResult(0)
-      if (
-        chunk.status === 'queued' ||
-        chunk.status === 'enqueue_failed' ||
-        (chunk.status === 'processing' && chunk.updatedAt === values[2])
-      ) {
-        chunk.status = 'processing'
-        chunk.updatedAt = String(values[0])
-        return runResult(1)
-      }
-      return runResult(0)
-    }
-    if (query.includes("UPDATE email_job_chunks SET status = 'sent'")) {
-      if (this.failSentUpdate) {
-        throw new Error('D1 unavailable')
-      }
-      const chunk = this.chunks.get(String(values[1]))
-      if (chunk) {
-        chunk.status = 'sent'
-        chunk.updatedAt = String(values[0])
-      }
-      return runResult(chunk ? 1 : 0)
-    }
-    if (query.includes('UPDATE email_jobs') && query.includes("SET status = 'sent'")) {
-      const jobId = String(values[1])
-      const hasUnsentChunks = Array.from(this.chunks.values()).some(
-        (chunk) => chunk.jobId === jobId && chunk.status !== 'sent'
-      )
-      const job = this.jobs.get(jobId)
-      if (job && !hasUnsentChunks) {
-        job.status = 'sent'
-        return runResult(1)
-      }
-      return runResult(0)
-    }
-    if (query.includes("SET status = 'enqueue_failed'") && query.includes('email_job_chunks')) {
-      const chunk = this.chunks.get(String(values[2]))
-      if (chunk) chunk.status = 'enqueue_failed'
-      return runResult(chunk ? 1 : 0)
-    }
-    if (query.includes("SET status = 'enqueue_failed'") && query.includes('email_jobs')) {
-      const job = this.jobs.get(String(values[2]))
-      if (job) job.status = 'enqueue_failed'
-      return runResult(job ? 1 : 0)
-    }
-    return runResult(0)
-  }
-
-  async first<T>(query: string, values: unknown[]): Promise<T | null> {
-    if (query.includes('FROM email_jobs') && query.includes('JOIN email_job_chunks')) {
-      const job = this.jobs.get(String(values[0]))
-      const chunk = this.chunks.get(String(values[1]))
-      if (!job || !chunk) return null
-      return {
-        job_status: job.status,
-        chunk_status: chunk.status,
-        chunk_updated_at: chunk.updatedAt
-      } as T
-    }
-    return null
-  }
-}
-
-function runResult(changes: number): D1RunResult {
-  return {
-    success: true,
-    results: [],
-    meta: { changes }
-  }
-}
-
-vi.mock('../templates', () => ({
-  renderTemplate: vi.fn()
-}))
-
-import handler from '../index'
-import { renderTemplate } from '../templates'
 
 function createMessageBatch(messages: MessageLike[]) {
   return {
@@ -162,7 +31,7 @@ function createEnv(emailSend = vi.fn().mockResolvedValue({ messageId: 'msg-1' })
   }
 }
 
-describe('email-consumer queue handler', () => {
+describe('email queue consumer', () => {
   beforeEach(() => {
     vi.clearAllMocks()
     vi.mocked(renderTemplate).mockResolvedValue({ html: '<h1>Test</h1>', text: 'Test' })
@@ -193,7 +62,7 @@ describe('email-consumer queue handler', () => {
       }
     ])
 
-    await handler.queue(batch as never, createEnv(emailSend) as never)
+    await queueHandler(batch as never, createEnv(emailSend) as never)
     expect(emailSend).toHaveBeenCalledWith({
       to: ['a@example.com', 'b@example.com'],
       from: 'sender@example.com',
@@ -229,7 +98,7 @@ describe('email-consumer queue handler', () => {
       }
     ])
 
-    await handler.queue(batch as never, createEnv(emailSend) as never)
+    await queueHandler(batch as never, createEnv(emailSend) as never)
     expect(retry).toHaveBeenCalled()
     expect(ack).not.toHaveBeenCalled()
   })
@@ -258,7 +127,7 @@ describe('email-consumer queue handler', () => {
       }
     ])
 
-    await handler.queue(batch as never, createEnv() as never)
+    await queueHandler(batch as never, createEnv() as never)
     expect(ack).toHaveBeenCalled()
     expect(vi.mocked(renderTemplate)).not.toHaveBeenCalled()
   })
@@ -278,7 +147,7 @@ describe('email-consumer queue handler', () => {
       }
     ])
 
-    await handler.queue(batch as never, createEnv() as never)
+    await queueHandler(batch as never, createEnv() as never)
     expect(ack).toHaveBeenCalled()
     expect(retry).not.toHaveBeenCalled()
     expect(vi.mocked(renderTemplate)).not.toHaveBeenCalled()
@@ -312,7 +181,7 @@ describe('email-consumer queue handler', () => {
       }
     ])
 
-    await handler.queue(batch as never, createEnv(emailSend, db) as never)
+    await queueHandler(batch as never, createEnv(emailSend, db) as never)
     expect(emailSend).not.toHaveBeenCalled()
     expect(ack).toHaveBeenCalled()
     expect(retry).not.toHaveBeenCalled()
@@ -345,7 +214,7 @@ describe('email-consumer queue handler', () => {
       }
     ])
 
-    await handler.queue(batch as never, createEnv(emailSend, db) as never)
+    await queueHandler(batch as never, createEnv(emailSend, db) as never)
     expect(emailSend).toHaveBeenCalledTimes(1)
     expect(ack).toHaveBeenCalled()
     expect(retry).not.toHaveBeenCalled()
