+++ packages/email/src/tests/enqueue.test.ts
@@ -1,94 +1,6 @@
 import { describe, it, expect, vi } from 'vitest'
-import app from '../index'
-
-interface D1RunResult {
-  meta: { changes: number }
-  results: unknown[]
-  success: true
-}
-
-class TestD1Statement {
-  private values: unknown[] = []
-
-  constructor(
-    private readonly query: string,
-    private readonly db: TestD1Database
-  ) {}
-
-  bind(...values: unknown[]): TestD1Statement {
-    this.values = values
-    return this
-  }
-
-  async run(): Promise<D1RunResult> {
-    return this.db.run(this.query, this.values)
-  }
-
-  async first<T>(): Promise<T | null> {
-    return this.db.first<T>(this.query, this.values)
-  }
-}
-
-class TestD1Database {
-  readonly jobs = new Map<string, { status: string; chunkCount: number }>()
-  readonly chunks = new Map<string, { jobId: string; status: string }>()
-
-  prepare(query: string): TestD1Statement {
-    return new TestD1Statement(query, this)
-  }
-
-  async batch(statements: TestD1Statement[]): Promise<D1RunResult[]> {
-    return Promise.all(statements.map((statement) => statement.run()))
-  }
-
-  async run(query: string, values: unknown[]): Promise<D1RunResult> {
-    if (query.includes('INSERT INTO email_jobs')) {
-      const jobId = String(values[0])
-      if (this.jobs.has(jobId)) {
-        throw new Error('SQLITE_CONSTRAINT: UNIQUE constraint failed: email_jobs.job_id')
-      }
-      this.jobs.set(jobId, { status: 'queued', chunkCount: Number(values[5]) })
-      return runResult(1)
-    }
-    if (query.includes('INSERT INTO email_job_chunks')) {
-      const messageId = String(values[0])
-      this.chunks.set(messageId, { jobId: String(values[1]), status: 'queued' })
-      return runResult(1)
-    }
-    if (query.includes("UPDATE email_jobs SET status = 'queued'")) {
-      const job = this.jobs.get(String(values[1]))
-      if (job) job.status = 'queued'
-      return runResult(job ? 1 : 0)
-    }
-    if (query.includes("UPDATE email_jobs SET status = 'enqueue_failed'")) {
-      const job = this.jobs.get(String(values[2]))
-      if (job) job.status = 'enqueue_failed'
-      return runResult(job ? 1 : 0)
-    }
-    if (query.includes("UPDATE email_job_chunks SET status = 'enqueue_failed'")) {
-      const chunk = this.chunks.get(String(values[2]))
-      if (chunk) chunk.status = 'enqueue_failed'
-      return runResult(chunk ? 1 : 0)
-    }
-    return runResult(0)
-  }
-
-  async first<T>(query: string, values: unknown[]): Promise<T | null> {
-    if (query.includes('SELECT status FROM email_jobs')) {
-      const job = this.jobs.get(String(values[0]))
-      return (job ? { status: job.status } : null) as T | null
-    }
-    return null
-  }
-}
-
-function runResult(changes: number): D1RunResult {
-  return {
-    success: true,
-    results: [],
-    meta: { changes }
-  }
-}
+import { app } from '../enqueue'
+import { TestD1Database } from './helpers/d1'
 
 function createTestEnv(authToken = 'test-token') {
   return {
