+++ packages/email/src/tests/templates.test.ts
