+++ packages/email/tsconfig.json
